🚚 rename: pages 파일 rename
