+++ src/pages/FindPassword/index.tsx
