--- src/pages/Post/index.ts
@@ -1,3 +1,0 @@
-export { default as PostCreatePage } from "./PostCreatePage";
-export { default as PostDetailPage } from "./PostDetailPage";
-export { default as PostUpdatePage } from "./PostUpdatePage";
