--- src/pages/index.ts
@@ -1,11 +1,0 @@
-export { default as BookmarkPage } from "./Bookmark/BookMarkPage";
-export { default as FindPasswordPage } from "./FindPassword/FindPasswordPage";
-export { default as LikePage } from "./Like/LikePage";
-export { default as LogInPage } from "./LogIn/LogInPage";
-export { default as MainPage } from "./Main/MainPage";
-export { PostCreatePage, PostDetailPage, PostUpdatePage } from "./Post";
-export { default as SearchPage } from "./Search/SearchPage";
-export { default as SignUpPage } from "./SignUp/SignUpPage";
-export { default as UserPage } from "./User/UserPage";
-export { default as UserProfileCreatePage } from "./User/UserProfileCreatePage";
-export { default as UserProfileUpdatePage } from "./User/UserProfileUpdatePage";
